Added a compiler memory barrier between sub-blocks to ensure that none of the writes into what was S2 during processing of the previous sub-block are postponed until after a read from S0 or S1 in the inline asm code for the current sub-block.  This potential problem was never observed so far due to other constraints that we have(*), but strictly speaking those constraints were insufficient to guarantee it can't occur.
(*) The registers to be written into S2 are overwritten by the last pwxform
round, so their old values would need to be preserved over that round in order
to expose the problem.  This would involve generating more and likely slower
code, thus not a lucrative "optimization" for a compiler to perform.

CVS-ID: yescrypt-simd.c 1.11